finished support for general adding of variables.
Mostly everything works. i need to make it split on pluses only if there are unresolved paranthesis, do division, and figure out variables to a variable power.  please note if development continues, input of exponents will need to be wrapped in paranthesis, so it can determine if it is to split on pluses. ex: 3x^(1+2) instead of 3x^1+2