implemented use of self, see description for errors
I implemented the use of self, so the class is actually being taken advantage of. I deleted decimal_to_fraction, as it wasnt needed. For the trig functions and calculation, I have not implemented the alternative method where it takes an input, and returns it instead of changing the object variable. The trig function do not work, nor does paranthesis because of this. It should be a simple fix.