Optimized calculation(), added trig functionality back (mostly)
I added an object variable message_functs, that is a 2d list with the message to print and the function it is correlated with. This is used in the calculation function, when calling the stage functions.

I also added functionality to the sin, cos, and tan functions because they previously relied on having an alternate input that returns a value. I am not sure if this is the best practice, but it works.